+++ CHANGELOG.md
@@ -3,7 +3,14 @@ All notable changes to this project will be documented in this file.
 Read `release_notes.md` for commit level details.
 
 ## [Unreleased]
+
+This release has a breaking change about an implicit wait.
+Ruby client sets `0` seconds as implicit wait by default from this release since it is the default spec behaviour in WebDriver while Ruby client had set `20` seconds for it.
+
 ### Enhancements
+- Breaking changes
+    - Set implicit wait zero by default
+        - Can configure `wait: 20` as `appium_lib` capability to keep the behaviour
 - [Experimental] Add `direct_connect` capability for the Ruby client in order to handle `directConnect` capability in a create session response by Appium server
     - Update http client following `directConnectProtocol`, `directConnectHost`, `directConnectPort` and `directConnectPath`
       if `direct_connect` capability for ruby_lib_core is `true`